# oyster-hub/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oyster Hub API",
    "version": "0.1.0",
    "description": "Driver-backed workspace provisioning, aggregation, and scoped proxy API for Oyster servers. The prototype driver is llmbox."
  },
  "servers": [{ "url": "/api/v1" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/environments": {
      "get": {
        "summary": "List driver-discovered and cloud environments",
        "responses": { "200": { "description": "Environment list", "content": { "application/json": { "schema": { "type": "object", "required": ["environments"], "properties": { "environments": { "type": "array", "items": { "$ref": "#/components/schemas/Environment" } } } } } } } }
      },
      "post": {
        "summary": "Provision a cloud VM as a reverse-connected Oyster box",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateEnvironment" } } } },
        "responses": { "201": { "description": "VM provisioning started" }, "400": { "description": "Invalid request" }, "409": { "description": "Provider credentials are not configured" }, "502": { "description": "Provider request failed" } }
      }
    },
...
    "/workspaces": {
      "get": {
        "summary": "Query the configured driver and list discovered workspaces",
        "responses": { "200": { "description": "Workspace list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkspaceList" } } } } }
      },
      "post": {
        "summary": "Create a workspace through the configured driver",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkspace" } } } },
        "responses": {
          "201": { "description": "Workspace created", "content": { "application/json": { "schema": { "type": "object", "required": ["workspace"], "properties": { "workspace": { "$ref": "#/components/schemas/Workspace" } } } } } },
          "400": { "description": "Invalid workspace request" },
          "405": { "description": "Selected driver cannot create workspaces" },
          "502": { "description": "Workspace driver failure" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" }
    },
...
    "schemas": {
...
      "Environment": {
        "type": "object",
        "required": ["id", "name", "status"],
        "properties": {
          "id": { "type": "string", "description": "Driver or cloud environment ID" },
          "name": { "type": "string" },
          "status": { "enum": ["provisioning", "awaiting_agent", "online", "offline", "failed"] },
          "default": { "type": "boolean" },
          "local": { "type": "boolean" },
          "cloud": { "type": "boolean" },
          "provider": { "type": "object", "additionalProperties": true }
        }
      },
      "Workspace": {
        "type": "object",
        "required": ["environmentId", "environmentName", "id", "name", "url"],
        "properties": {
          "environmentId": { "type": "string", "description": "Physical or cloud device ID" },
          "environmentName": { "type": "string", "description": "Physical or cloud device label" },
          "id": { "type": "string", "description": "Project microVM ID" },
          "name": { "type": "string" },
          "url": { "type": ["string", "null"], "format": "uri" },
          "status": { "enum": ["online", "offline", "provisioning"] },
          "latencyMs": { "type": ["integer", "null"] },
          "provider": { "type": "object", "additionalProperties": true },
          "results": { "type": "object", "additionalProperties": true },
          "errors": { "type": "object", "additionalProperties": { "type": "string" } }
        }
      },
      "WorkspaceList": { "type": "object", "required": ["driver", "workspaces"], "properties": { "driver": { "type": "object", "required": ["type", "endpoint", "capabilities"], "properties": { "type": { "type": "string" }, "endpoint": { "type": "string", "format": "uri" }, "capabilities": { "type": "object", "required": ["list", "create", "remove"], "properties": { "list": { "type": "boolean" }, "create": { "type": "boolean" }, "remove": { "type": "boolean" } } } } }, "workspaces": { "type": "array", "items": { "$ref": "#/components/schemas/Workspace" } } } },
      "CreateEnvironment": {
        "type": "object",
        "required": ["provider", "name", "region", "size", "image"],
        "properties": {
          "provider": { "enum": ["digitalocean", "aws", "gcp"] },
          "name": { "type": "string", "pattern": "^[A-Za-z](?:[A-Za-z0-9-]{0,61}[A-Za-z0-9])?$" },
          "region": { "type": "string" },
          "size": { "type": "string" },
          "image": { "type": "string" }
        }
      },
      "CreateWorkspace": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string", "pattern": "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$" },
          "name": { "type": "string", "maxLength": 500 },
          "spoke": { "type": "string", "maxLength": 100 },
          "diskBytes": { "type": "integer", "minimum": 0 }
        }
      },
...
    }
  }
}